+++ backend/src/main/java/com/carrot/application/like/domain/PostLike.java
@@ -1,7 +1,7 @@
 package com.carrot.application.like.domain;
 
 import com.carrot.application.common.BaseEntity;
-import com.carrot.application.post.domain.Post;
+import com.carrot.application.post.domain.entity.Post;
 import com.carrot.application.user.domain.User;
 import lombok.Builder;
 import lombok.Getter;
