+++ backend/src/main/java/com/carrot/application/post/repository/PostImageRepository.java
@@ -1,6 +1,6 @@
 package com.carrot.application.post.repository;
 
-import com.carrot.application.post.domain.PostImage;
+import com.carrot.application.post.domain.entity.PostImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostImageRepository extends JpaRepository<PostImage, Long> {
