+++ backend/src/main/java/com/carrot/application/post/repository/PostRepository.java
@@ -1,8 +1,7 @@
 package com.carrot.application.post.repository;
 
-import com.carrot.application.post.domain.Post;
+import com.carrot.application.post.domain.entity.Post;
 import com.carrot.global.error.CarrotRuntimeException;
-import com.carrot.global.error.ErrorCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import static com.carrot.global.error.ErrorCode.*;
