+++ backend/src/main/java/com/carrot/application/post/service/PostWriteService.java
@@ -1,7 +1,7 @@
 package com.carrot.application.post.service;
 
-import com.carrot.application.post.domain.Post;
-import com.carrot.application.post.domain.PostImage;
+import com.carrot.application.post.domain.entity.Post;
+import com.carrot.application.post.domain.entity.PostImage;
 import com.carrot.application.post.repository.PostRepository;
 import com.carrot.application.region.domain.Region;
 import com.carrot.application.region.repository.RegionRepository;
